Update FedEx API integration and refactor related classes
This commit includes a major update to the FedEx API integration and a significant refactoring of related classes. The `Newtonsoft.Json` package was updated, and several interfaces and their implementations were removed, including `IFedExAuthClient` and `FedExAuthClient`. The `FedExServiceCollectionExtensions` class was updated, replacing the `AddFedExAuthApiClient` method with `AddFedExApiClients`. Several OpenApi references were added in `EasyKeys.Shipping.FedEx.Abstractions.csproj`, and the `AuthRequestMiddleware` class was updated to use `IFedexApiAuthenticatorService`. A new `AuthorizationApi.cs` file was added, and changes were made to the `authorization.json` file. The `JsonSerializerOptions` property in `FedExApiOptions` class was replaced with `JsonSerializerSettings`, and a new namespace `EasyKeys.Shipping.FedEx.Abstractions.Services` was added. Several classes related to address validation and rates and transit times were removed. The `FedexRateProvider.cs` file was significantly refactored, and a new method `AddRestApiFedExRateProvider` was added to the `FedExShippingServiceCollectionExtensions` class. A new class `FedExShipmentProvider` was added, and changes were made to the namespaces of several classes. Changes were also made to `FedExRateProviderTests.cs` and `FedExShipmentProvider.cs`.